wip: split time input component
